Restructure of function foleder
