+++ components/calculator/functions/isZeroish.ts
